refactor: consolidate task data fetching and state management in TaskTabContent
